+++ src/AudioInPort.cpp
@@ -27,16 +27,16 @@ AudioInPort::AudioInPort(int position) {
 
 AudioInPort::~AudioInPort() = default;
 
-void AudioInPort::configure(float scale, float offset) {
+void AudioInPort::configure(int32_t scale, int32_t offset) {
    setScale(scale);
    setOffset(offset);
 }
 
-float IRAM_ATTR audioinport_read(signal_t *handle, uint64_t time) {
+int32_t IRAM_ATTR audioinport_read(signal_t *handle, uint64_t time) {
    auto *port = (audio_in_port_t *) handle->data;
    if( time > port->lastCalc ) {
       port->lastCalc = time;
-      float result = Demiurge::runtime().inputs()[port->position];
+      int32_t result = Demiurge::runtime().inputs()[port->position];
       port->cached = result;
       return result;
    }
