+++ src/CvInPort.h
@@ -26,10 +26,10 @@ See the License for the specific language governing permissions and
 typedef struct {
    int position;
    uint64_t lastCalc;
-   float cached;
+   int32_t cached;
 } cv_in_port_t;
 
-float cvinport_read(signal_t *handle, uint64_t time);
+int32_t cvinport_read(signal_t *handle, uint64_t time);
 
 class CvInPort : public Signal {
