+++ src/Inverter.h
@@ -25,10 +25,10 @@ typedef struct {
    signal_t *scale;
    signal_t *midpoint;
    uint64_t lastCalc;
-   float cached;
+   int32_t cached;
 } inverter_t;
 
-float inverter_read(signal_t *handle, uint64_t time);
+int32_t inverter_read(signal_t *handle, uint64_t time);
 
 class Inverter : public Signal {
