+++ src/Mixer.h
@@ -26,10 +26,10 @@ See the License for the specific language governing permissions and
 typedef struct {
    signal_t *inputs[DEMIURGE_MAX_MIXER_IN];
    uint64_t lastCalc;
-   float cached;
+   int32_t cached;
 } mixer_t;
 
-float mixer_read(signal_t *handle, uint64_t time);
+int32_t mixer_read(signal_t *handle, uint64_t time);
 
 
 class Mixer : public Signal {
