+++ src/Pan.h
@@ -20,15 +20,15 @@ See the License for the specific language governing permissions and
 #include "Signal.h"
 
 typedef struct {
-   float factor;
+   int32_t factor;
    signal_t *me;
    signal_t *control;
    signal_t *hostInput;
    uint64_t lastCalc;
-   float cached;
+   int32_t cached;
 } panchannel_t;
 
-float IRAM_ATTR panchannel_read(signal_t *handle, uint64_t time);
+int32_t IRAM_ATTR panchannel_read(signal_t *handle, uint64_t time);
 
 class PanChannel;
 
@@ -57,7 +57,7 @@ private :
 class PanChannel : public Signal {
 
 public:
-   PanChannel(Pan *host, float factor);
+   PanChannel(Pan *host, int32_t factor);
 
    ~PanChannel() override;
