+++ src/Passthru.h
@@ -24,10 +24,10 @@ typedef struct {
    signal_t *me;
    signal_t *input;
    uint64_t lastCalc;
-   float cached;
+   int32_t cached;
 } passthru_t;
 
-float passthru_read(signal_t *handle, uint64_t time);
+int32_t passthru_read(signal_t *handle, uint64_t time);
 
 class Passthru : public Signal {
