+++ src/Potentiometer.h
@@ -26,10 +26,10 @@ See the License for the specific language governing permissions and
 typedef struct {
    int position;
    uint64_t lastCalc;
-   float cached;
+   int32_t cached;
 } potentiometer_t;
 
-float potentiometer_read(signal_t *handle, uint64_t time);
+int32_t potentiometer_read(signal_t *handle, uint64_t time);
 
 
 class Potentiometer : public Signal {
