+++ src/Volume.h
@@ -25,10 +25,10 @@ typedef struct {
    signal_t *input;
    signal_t *control;
    uint64_t lastCalc;
-   float cached;
+   int32_t cached;
 } volume_t;
 
-float volume_read(signal_t *handle, uint64_t time);
+int32_t volume_read(signal_t *handle, uint64_t time);
 
 class Volume : public Signal {
